requirements.txt
litellm
fastapi
watchdog
pytest
beautifulsoup4
trafilatura
requests
duckdb
sqlalchemy
google-api-python-client
